+++ mopidy/backends/spotify/translator.py
@@ -1,11 +1,15 @@
 import datetime as dt
+import logging
 
-from spotify import Link
+from spotify import Link, SpotifyError
 
+from mopidy import settings
+from mopidy.backends.spotify import ENCODING
 from mopidy.models import Artist, Album, Track, Playlist
-from mopidy.backends.libspotify import ENCODING
 
-class LibspotifyTranslator(object):
+logger = logging.getLogger('mopidy.backends.spotify.translator')
+
+class SpotifyTranslator(object):
     @classmethod
     def to_mopidy_artist(cls, spotify_artist):
         if not spotify_artist.is_loaded():
@@ -39,15 +43,22 @@ class LibspotifyTranslator(object):
             track_no=spotify_track.index(),
             date=date,
             length=spotify_track.duration(),
-            bitrate=160,
+            bitrate=(settings.SPOTIFY_HIGH_BITRATE and 320 or 160),
         )
 
     @classmethod
     def to_mopidy_playlist(cls, spotify_playlist):
         if not spotify_playlist.is_loaded():
             return Playlist(name=u'[loading...]')
-        return Playlist(
-            uri=str(Link.from_playlist(spotify_playlist)),
-            name=spotify_playlist.name().decode(ENCODING),
-            tracks=[cls.to_mopidy_track(t) for t in spotify_playlist],
-        )
+        # FIXME Replace this try-except with a check on the playlist type,
+        # which is currently not supported by pyspotify, to avoid handling
+        # playlist folder boundaries like normal playlists.
+        try:
+            return Playlist(
+                uri=str(Link.from_playlist(spotify_playlist)),
+                name=spotify_playlist.name().decode(ENCODING),
+                tracks=[cls.to_mopidy_track(t) for t in spotify_playlist],
+            )
+        except SpotifyError, e:
+            logger.warning(u'Failed translating Spotify playlist '
+                '(probably a playlist folder boundary): %s', e)